% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{previewbw}
\alias{previewbw}
\title{Preview a `ggplot` in black and white}
\usage{
previewbw(plot)
}
\arguments{
\item{plot}{Plot that you want to view in black and white. If it's not speficied
then the last plot is used. Only works with plots made with `ggplot2` (and
extensions).}
}
\description{
Preview a `ggplot` in black and white
}
